requirements.txt
Flask==3.0.0
langchain>=1.0.0
langchain-openai>=0.0.5
langchain-core>=1.0.0
langchain-community>=0.4.0
httpx==0.26.0
python-dotenv==1.0.0
